Refactor connected sites
